refactor: move Messenger files to `apps/messenger/` Clears up the `src` directory Is consistent with putting all apps in `apps/`
