+++ src/apps/app-router.tsx
@@ -5,7 +5,7 @@
  */
 
 import { Redirect, Route, Switch } from 'react-router-dom';
-import { MessengerMain } from '../messenger-main';
+import { MessengerApp } from './messenger';
 import { ExplorerApp } from './explorer';
 
 const redirectToRoot = () => <Redirect to={'/'} />;
@@ -13,8 +13,8 @@ const redirectToRoot = () => <Redirect to={'/'} />;
 export const AppRouter = () => {
   return (
     <Switch>
-      <Route path='/conversation/:conversationId' exact component={MessengerMain} />
-      <Route path='/' exact component={MessengerMain} />
+      <Route path='/conversation/:conversationId' exact component={MessengerApp} />
+      <Route path='/' exact component={MessengerApp} />
       <Route path='/explorer' component={ExplorerApp} />
       <Route component={redirectToRoot} />
     </Switch>
