+++ src/apps/app-router.vitest.tsx
@@ -4,9 +4,9 @@ import { AppRouter } from './app-router';
 import { vi } from 'vitest';
 import { renderWithProviders } from '../test-utils';
 
-vi.mock('../messenger-main', () => ({
-  MessengerMain: () => {
-    return <div data-testid='messenger-main' />;
+vi.mock('./messenger', () => ({
+  MessengerApp: () => {
+    return <div data-testid='messenger-app' />;
   },
 }));
 
@@ -21,16 +21,16 @@ const renderComponent = (route: string | undefined = '/') => {
 describe(AppRouter, () => {
   it('should render MessengerMain component when route is /', () => {
     renderComponent('/');
-    expect(screen.getByTestId('messenger-main')).toBeTruthy();
+    expect(screen.getByTestId('messenger-app')).toBeTruthy();
   });
 
   it('should render MessengerMain component when route is /conversation/:conversationId', () => {
     renderComponent('/conversation/123');
-    expect(screen.getByTestId('messenger-main')).toBeTruthy();
+    expect(screen.getByTestId('messenger-app')).toBeTruthy();
   });
 
   it('should redirect to / when route is invalid', () => {
     renderComponent('/foo-bar');
-    expect(screen.getByTestId('messenger-main')).toBeTruthy();
+    expect(screen.getByTestId('messenger-app')).toBeTruthy();
   });
 });
