+++ src/apps/messenger/index.tsx
@@ -1,0 +1,2 @@
+export { MessengerMain as MessengerApp } from './messenger-main';
+export type { Properties as MessengerProperties } from './messenger-main';
